Добавляет модульные файлы js
