+++ js/variables.js
@@ -1,0 +1,49 @@
+export const POSTS_COUNT = 25;
+export const AVATARS_COUNT = 6;
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
+export const COMMENTS_COUNT = 30;
+
+export const POSTS_DESCRIPTIONS = [
+  'И мое сердце остановилось, мое сердце замерло.',
+  'Не слушай никого, лишь свое сердце',
+  'Как мало нужно для счастья.',
+  'Теплые воспоминания в холодное время года.',
+  'Всегда готовы покорять новые вершины.',
+  'Настойчивость окупается сполна.',
+  'Будь голосом, а не эхом.',
+  'Отсутствие сна — это не проблема. Проблема, когда ты не знаешь, ради чего просыпаешься по утрам.',
+  'Временно в режиме off-line.',
+  'Нормальные люди по утрам просыпаются, а я восстаю.',
+];
+
+export const POSTS_COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const COMMENT_NAMES = [
+  'Серёжа',
+  'Олег',
+  'Оля',
+  'Саша',
+  'Люся',
+  'Женя',
+  'Дима',
+  'Данила',
+  'Артём',
+  'Настя',
+  'Лиза',
+  'Миша',
+  'Роберт',
+  'Паша',
+  'Ляйсан',
+  'Гарик',
+  'Антон',
+  'Арсений',
+  'Ксюша',
+];
